split internal dial options
