--- internal/balancer/balancer.go
@@ -1,34 +1,0 @@
-/*
- * Copyright 2018 gRPC authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-// Package balancer contains gRPC-internal code for load balancing, to avoid
-// polluting the godoc of the top-level grpc package.
-package balancer // import "google.golang.org/grpc/internal/balancer"
-
-import (
-	"sync"
-
-  _ "google.golang.org/grpc"
-	_ "google.golang.org/grpc/connectivity"
-)
-
-// set by clientconn.go
-var (
-  WithContextDialer interface{} // func(context.Context, string) (net.Conn, error) grpc.DialOption
-  WithResolverBuilder interface{} // func (resolver.Builder) grpc.DialOption
-)
-
