+++ internal/internal.go
@@ -15,8 +15,9 @@
  *
  */
 
-// Package internal contains gRPC-internal code for testing, to avoid polluting
-// the godoc of the top-level grpc package.
+// Package internal contains gRPC-internal code, to avoid polluting
+// the godoc of the top-level grpc package.  It must not import any grpc 
+// symbols to avoid circular dependencies.
 package internal
 
 // TestingUseHandlerImpl enables the http.Handler-based server implementation.
@@ -25,3 +26,9 @@ package internal
 // The provided grpcServer must be of type *grpc.Server. It is untyped
 // for circular dependency reasons.
 var TestingUseHandlerImpl func(grpcServer interface{})
+
+// set by clientconn.go
+var (
+  WithContextDialer interface{} // func(context.Context, string) (net.Conn, error) grpc.DialOption
+  WithResolverBuilder interface{} // func (resolver.Builder) grpc.DialOption
+)
